refactor: directory structure and exports
