--- packages/sdk/src/External.ts
@@ -1,21 +1,0 @@
-export * as AaveV2 from "./External/AaveV2.js";
-export * as AaveV3 from "./External/AaveV3.js";
-export * as ArrakisV2 from "./External/ArrakisV2.js";
-export * as BalancerV2 from "./External/BalancerV2.js";
-export * as CompoundV2 from "./External/CompoundV2.js";
-export * as CompoundV3 from "./External/CompoundV3.js";
-export * as Convex from "./External/Convex.js";
-export * as Curve from "./External/Curve.js";
-export * as DelegateVotes from "./External/DelegateVotes.js";
-export * as Idle from "./External/Idle.js";
-export * as Kiln from "./External/Kiln.js";
-export * as Lido from "./External/Lido.js";
-export * as Liquity from "./External/Liquity.js";
-export * as Maker from "./External/Maker.js";
-export * as Maple from "./External/Maple.js";
-export * as Morpho from "./External/Morpho.js";
-export * as StakeWiseV3 from "./External/StakeWiseV3.js";
-export * as TheGraph from "./External/TheGraph.js";
-export * as UniswapV2 from "./External/UniswapV2.js";
-export * as UniswapV3 from "./External/UniswapV3.js";
-export * as Yearn from "./External/Yearn.js";
