--- packages/sdk/src/External/StakeWiseV3.ts
@@ -1,20 +1,0 @@
-import { type Address, type PublicClient, parseAbi } from "viem";
-import { getBalanceOf } from "../Assets.js";
-import { Viem } from "../Utils.js";
-
-export async function getStakedEthBalance(
-  client: PublicClient,
-  args: Viem.ContractCallParameters<{
-    account: Address;
-    stakeWiseVaultAddress: Address;
-  }>,
-) {
-  const sharesBalance = await getBalanceOf(client, { asset: args.stakeWiseVaultAddress, owner: args.account });
-
-  return Viem.readContract(client, args, {
-    abi: parseAbi(["function convertToAssets(uint256 _shares) view returns (uint256 assets_)"]),
-    functionName: "convertToAssets",
-    address: args.stakeWiseVaultAddress,
-    args: [sharesBalance],
-  });
-}
