--- packages/sdk/src/Portfolio/Extension.ts
@@ -1,1 +1,0 @@
-export { type CallOnExtensionParams } from "../_internal/Extensions.js";
