+++ packages/sdk/src/Portfolio/Integrations.ts
@@ -1,19 +1,24 @@
 export * as AaveV2 from "./Integrations/AaveV2.js";
 export * as AaveV3 from "./Integrations/AaveV3.js";
+export * as ArbitraryLoan from "./Integrations/ArbitraryLoan.js";
 export * as ArrakisV2 from "./Integrations/ArrakisV2.js";
-export * as BalancerV2Liquidity from "./Integrations/BalancerV2Liquidity.js";
+export * as BalancerV2 from "./Integrations/BalancerV2.js";
 export * as CompoundV2 from "./Integrations/CompoundV2.js";
 export * as CompoundV3 from "./Integrations/CompoundV3.js";
-export * as ConvexCurveLpStaking from "./Integrations/ConvexCurveLpStaking.js";
-export * as CurveExchange from "./Integrations/CurveExchange.js";
-export * as CurveLiquidity from "./Integrations/CurveLiquidity.js";
+export * as Convex from "./Integrations/Convex.js";
+export * as Curve from "./Integrations/Curve.js";
 export * as Erc4626 from "./Integrations/Erc4626.js";
 export * as IdleV4 from "./Integrations/IdleV4.js";
+export * as Kiln from "./Integrations/Kiln.js";
+export * as Lido from "./Integrations/Lido.js";
+export * as Liquity from "./Integrations/Liquity.js";
+export * as Maple from "./Integrations/Maple.js";
 export * as OneInchV5 from "./Integrations/OneInchV5.js";
 export * as ParaswapV5 from "./Integrations/ParaswapV5.js";
-export * as UniswapV2Exchange from "./Integrations/UniswapV2Exchange.js";
-export * as UniswapV2Liquidity from "./Integrations/UniswapV2Liquidity.js";
+export * as StakeWiseV3 from "./Integrations/StakeWiseV3.js";
+export * as TheGraphDelegation from "./Integrations/TheGraph.js";
+export * as UniswapV2 from "./Integrations/UniswapV2.js";
 export * as UniswapV3 from "./Integrations/UniswapV3.js";
-export * as YearnVaultV2 from "./Integrations/YearnVaultV2.js";
+export * as YearnV2 from "./Integrations/YearnV2.js";
 export * as ZeroExV2 from "./Integrations/ZeroExV2.js";
 export * as ZeroExV4 from "./Integrations/ZeroExV4.js";
