+++ packages/sdk/src/Portfolio/Integrations/ArrakisV2.ts
@@ -1,4 +1,5 @@
-import { type Address, type Hex, decodeAbiParameters, encodeAbiParameters } from "viem";
+import { type Address, type Hex, PublicClient, decodeAbiParameters, encodeAbiParameters } from "viem";
+import { Viem } from "../../Utils.js";
 import * as IntegrationManager from "../../_internal/IntegrationManager.js";
 
 //--------------------------------------------------------------------------------------------
@@ -82,5 +83,83 @@ export function redeemDecode(encoded: Hex): RedeemArgs {
     arrakisVault,
     sharesAmount,
     minIncomingUnderlyingAmounts,
+  };
+}
+
+//--------------------------------------------------------------------------------------------
+// THIRD PARTY READ FUNCTIONS
+//--------------------------------------------------------------------------------------------
+
+const resolverAbi = [
+  {
+    inputs: [
+      { internalType: "contract IArrakisV2", name: "vaultV2_", type: "address" },
+      { internalType: "uint256", name: "amount0Max_", type: "uint256" },
+      { internalType: "uint256", name: "amount1Max_", type: "uint256" },
+    ],
+    name: "getMintAmounts",
+    outputs: [
+      { internalType: "uint256", name: "amount0", type: "uint256" },
+      { internalType: "uint256", name: "amount1", type: "uint256" },
+      { internalType: "uint256", name: "mintAmount", type: "uint256" },
+    ],
+    stateMutability: "view",
+    type: "function",
+  },
+] as const;
+
+export async function getMintAmounts(
+  client: PublicClient,
+  args: Viem.ContractCallParameters<{
+    resolver: Address;
+    arrakisVault: Address;
+    amount0Max: bigint;
+    amount1Max: bigint;
+  }>,
+) {
+  const [amount0, amount1, mintAmount] = await Viem.readContract(client, args, {
+    abi: resolverAbi,
+    functionName: "getMintAmounts",
+    address: args.resolver,
+    args: [args.arrakisVault, args.amount0Max, args.amount1Max],
+  });
+
+  return {
+    amount0,
+    amount1,
+    mintAmount,
+  };
+}
+
+const helperAbi = [
+  {
+    inputs: [{ internalType: "contract IArrakisV2", name: "vault_", type: "address" }],
+    name: "totalUnderlying",
+    outputs: [
+      { internalType: "uint256", name: "amount0", type: "uint256" },
+      { internalType: "uint256", name: "amount1", type: "uint256" },
+    ],
+    stateMutability: "view",
+    type: "function",
+  },
+] as const;
+
+export async function totalUnderlying(
+  client: PublicClient,
+  args: Viem.ContractCallParameters<{
+    helper: Address;
+    arrakisVault: Address;
+  }>,
+) {
+  const [amount0, amount1] = await Viem.readContract(client, args, {
+    abi: helperAbi,
+    functionName: "totalUnderlying",
+    address: args.helper,
+    args: [args.arrakisVault],
+  });
+
+  return {
+    amount0,
+    amount1,
   };
 }
