+++ packages/sdk/src/Portfolio/Integrations/Convex.ts
@@ -1,9 +1,427 @@
 import * as Abis from "@enzymefinance/abis";
-import { type Address, type PublicClient, parseAbi } from "viem";
-import { Assertion, Viem } from "../Utils.js";
+import { type Address, type Hex, PublicClient, decodeAbiParameters, encodeAbiParameters, parseAbi } from "viem";
+import { Assertion, Viem } from "../../Utils.js";
+import * as ExternalPositionManager from "../../_internal/ExternalPositionManager.js";
+import * as IntegrationManager from "../../_internal/IntegrationManager.js";
+import { type RedeemType, isValidRedeemType } from "./Curve.js";
 
 //--------------------------------------------------------------------------------------------
-// CVX MINING
+// LEND AND STAKE
+//--------------------------------------------------------------------------------------------
+
+const lendAndStakeSelector = "0x29fa046e"; // lendAndStake(address,bytes,bytes)
+export const lendAndStake = IntegrationManager.makeUse(lendAndStakeSelector, lendAndStakeEncode);
+
+const lendAndStakeEncoding = [
+  {
+    name: "pool",
+    type: "address",
+  },
+  {
+    name: "orderedOutgoingAssetAmounts",
+    type: "uint256[]",
+  },
+  {
+    name: "incomingStakingToken",
+    type: "address",
+  },
+  {
+    name: "minIncomingStakingTokenAmount",
+    type: "uint256",
+  },
+  {
+    name: "useUnderlyings",
+    type: "bool",
+  },
+] as const;
+
+export type LendAndStakeArgs = {
+  pool: Address;
+  orderedOutgoingAssetAmounts: ReadonlyArray<bigint>;
+  incomingStakingToken: Address;
+  minIncomingStakingTokenAmount: bigint;
+  useUnderlyings: boolean;
+};
+
+export function lendAndStakeEncode(args: LendAndStakeArgs): Hex {
+  return encodeAbiParameters(lendAndStakeEncoding, [
+    args.pool,
+    args.orderedOutgoingAssetAmounts,
+    args.incomingStakingToken,
+    args.minIncomingStakingTokenAmount,
+    args.useUnderlyings,
+  ]);
+}
+
+export function lendAndStakeDecode(encoded: Hex): LendAndStakeArgs {
+  const [pool, orderedOutgoingAssetAmounts, incomingStakingToken, minIncomingStakingTokenAmount, useUnderlyings] =
+    decodeAbiParameters(lendAndStakeEncoding, encoded);
+
+  return {
+    pool,
+    orderedOutgoingAssetAmounts,
+    incomingStakingToken,
+    minIncomingStakingTokenAmount,
+    useUnderlyings,
+  };
+}
+
+//--------------------------------------------------------------------------------------------
+// CLAIM REWARDS
+//--------------------------------------------------------------------------------------------
+
+const claimRewardsSelector = "0xb9dfbacc"; // claimRewards(address,bytes,bytes)
+export const claimRewards = IntegrationManager.makeUse(claimRewardsSelector, claimRewardsEncode);
+
+const claimRewardsEncoding = [
+  {
+    name: "stakingToken",
+    type: "address",
+  },
+] as const;
+
+export type ClaimRewardsArgs = {
+  stakingToken: Address;
+};
+
+export function claimRewardsEncode(args: ClaimRewardsArgs): Hex {
+  return encodeAbiParameters(claimRewardsEncoding, [args.stakingToken]);
+}
+
+export function claimRewardsDecode(encoded: Hex): ClaimRewardsArgs {
+  const [stakingToken] = decodeAbiParameters(claimRewardsEncoding, encoded);
+
+  return { stakingToken };
+}
+
+//--------------------------------------------------------------------------------------------
+// STAKE
+//--------------------------------------------------------------------------------------------
+
+const stakeSelector = "0xfa7dd04d"; // stake(address,bytes,bytes)
+export const stake = IntegrationManager.makeUse(stakeSelector, stakeEncode);
+
+const stakeEncoding = [
+  {
+    name: "pool",
+    type: "address",
+  },
+  {
+    name: "incomingStakingToken",
+    type: "address",
+  },
+  {
+    name: "amount",
+    type: "uint256",
+  },
+] as const;
+
+export type StakeArgs = {
+  pool: Address;
+  incomingStakingToken: Address;
+  amount: bigint;
+};
+
+export function stakeEncode(args: StakeArgs): Hex {
+  return encodeAbiParameters(stakeEncoding, [args.pool, args.incomingStakingToken, args.amount]);
+}
+
+export function stakeDecode(encoded: Hex): StakeArgs {
+  const [pool, incomingStakingToken, amount] = decodeAbiParameters(stakeEncoding, encoded);
+
+  return { pool, incomingStakingToken, amount };
+}
+
+//--------------------------------------------------------------------------------------------
+// UNSTAKE
+//--------------------------------------------------------------------------------------------
+
+const unstakeSelector = "0x68e30677"; // unstake(address,bytes,bytes)
+export const unstake = IntegrationManager.makeUse(unstakeSelector, unstakeEncode);
+
+const unstakeEncoding = [
+  {
+    name: "pool",
+    type: "address",
+  },
+  {
+    name: "outgoingStakingToken",
+    type: "address",
+  },
+  {
+    name: "amount",
+    type: "uint256",
+  },
+] as const;
+
+export type UnstakeArgs = {
+  pool: Address;
+  outgoingStakingToken: Address;
+  amount: bigint;
+};
+
+export function unstakeEncode(args: UnstakeArgs): Hex {
+  return encodeAbiParameters(unstakeEncoding, [args.pool, args.outgoingStakingToken, args.amount]);
+}
+
+export function unstakeDecode(encoded: Hex): UnstakeArgs {
+  const [pool, outgoingStakingToken, amount] = decodeAbiParameters(unstakeEncoding, encoded);
+
+  return { pool, outgoingStakingToken, amount };
+}
+
+//--------------------------------------------------------------------------------------------
+// UNSTAKE AND REDEEM
+//--------------------------------------------------------------------------------------------
+
+const unstakeAndRedeemSelector = "0x8334eb99"; // unstakeAndRedeem(address,bytes,bytes)
+export const unstakeAndRedeem = IntegrationManager.makeUse(unstakeAndRedeemSelector, unstakeAndRedeemEncode);
+
+const unstakeAndRedeemEncoding = [
+  {
+    name: "pool",
+    type: "address",
+  },
+  {
+    name: "outgoingStakingToken",
+    type: "address",
+  },
+  {
+    name: "outgoingStakingTokenAmount",
+    type: "uint256",
+  },
+  {
+    name: "useUnderlyings",
+    type: "bool",
+  },
+  {
+    name: "redeemType",
+    type: "uint8",
+  },
+  {
+    name: "incomingAssetsData",
+    type: "bytes",
+  },
+] as const;
+
+export type UnstakeAndRedeemArgs = {
+  pool: Address;
+  outgoingStakingToken: Address;
+  outgoingStakingTokenAmount: bigint;
+  useUnderlyings: boolean;
+  redeemType: RedeemType;
+  incomingAssetsData: Hex;
+};
+
+export function unstakeAndRedeemEncode(args: UnstakeAndRedeemArgs): Hex {
+  return encodeAbiParameters(unstakeAndRedeemEncoding, [
+    args.pool,
+    args.outgoingStakingToken,
+    args.outgoingStakingTokenAmount,
+    args.useUnderlyings,
+    args.redeemType,
+    args.incomingAssetsData,
+  ]);
+}
+
+export function unstakeAndRedeemDecode(encoded: Hex): UnstakeAndRedeemArgs {
+  const [pool, outgoingStakingToken, outgoingStakingTokenAmount, useUnderlyings, redeemType, incomingAssetsData] =
+    decodeAbiParameters(unstakeAndRedeemEncoding, encoded);
+
+  if (!isValidRedeemType(redeemType)) {
+    Assertion.invariant(false, "Invalid redeem type");
+  }
+
+  return {
+    pool,
+    outgoingStakingToken,
+    outgoingStakingTokenAmount,
+    useUnderlyings,
+    redeemType,
+    incomingAssetsData,
+  };
+}
+
+//--------------------------------------------------------------------------------------------
+// EXTERNAL POSITION
+//--------------------------------------------------------------------------------------------
+
+export type Action = typeof Action[keyof typeof Action];
+export const Action = {
+  Lock: 0n,
+  Relock: 1n,
+  Withdraw: 2n,
+  ClaimRewards: 3n,
+  Delegate: 4n,
+} as const;
+
+export const create = ExternalPositionManager.createOnly;
+
+//--------------------------------------------------------------------------------------------
+// LOCK
+//--------------------------------------------------------------------------------------------
+
+export const lock = ExternalPositionManager.makeUse(Action.Lock, lockEncode);
+export const createAndLock = ExternalPositionManager.makeCreateAndUse(Action.Lock, lockEncode);
+
+const lockEncoding = [
+  {
+    name: "amount",
+    type: "uint256",
+  },
+  {
+    name: "spendRatio",
+    type: "uint256",
+  },
+] as const;
+
+export type LockArgs = {
+  amount: bigint;
+  spendRatio: bigint;
+};
+
+export function lockEncode(args: LockArgs): Hex {
+  return encodeAbiParameters(lockEncoding, [args.amount, args.spendRatio]);
+}
+
+export function lockDecode(encoded: Hex): LockArgs {
+  const [amount, spendRatio] = decodeAbiParameters(lockEncoding, encoded);
+
+  return {
+    amount,
+    spendRatio,
+  };
+}
+
+//--------------------------------------------------------------------------------------------
+// RELOCK
+//--------------------------------------------------------------------------------------------
+
+export const relock = ExternalPositionManager.makeUse(Action.Relock);
+
+//--------------------------------------------------------------------------------------------
+// WITHDRAW
+//--------------------------------------------------------------------------------------------
+
+export const withdraw = ExternalPositionManager.makeUse(Action.Withdraw);
+
+//--------------------------------------------------------------------------------------------
+// CLAIM REWARDS
+//--------------------------------------------------------------------------------------------
+
+export const claimVotingRewards = ExternalPositionManager.makeUse(Action.ClaimRewards, claimVotingRewardsEncode);
+
+const claimVotingRewardsEncoding = [
+  {
+    name: "allTokensToTransfer",
+    type: "address[]",
+  },
+  {
+    name: "claimLockerRewards",
+    type: "bool",
+  },
+  {
+    name: "extraRewardTokens",
+    type: "address[]",
+  },
+  {
+    components: [
+      {
+        name: "token",
+        type: "address",
+      },
+      {
+        name: "index",
+        type: "uint256",
+      },
+      {
+        name: "amount",
+        type: "uint256",
+      },
+      {
+        name: "merkleProof",
+        type: "bytes32[]",
+      },
+    ],
+    name: "votiumClaims",
+    type: "tuple[]",
+  },
+  {
+    name: "unstakeCvxCrv",
+    type: "bool",
+  },
+] as const;
+
+export type ClaimVotingRewardsArgs = {
+  allTokensToTransfer: ReadonlyArray<Address>;
+  claimLockerRewards: boolean;
+  extraRewardTokens: ReadonlyArray<Address>;
+  votiumClaims: ReadonlyArray<{
+    token: Address;
+    index: bigint;
+    amount: bigint;
+    merkleProof: ReadonlyArray<Hex>;
+  }>;
+  unstakeCvxCrv: boolean;
+};
+
+export function claimVotingRewardsEncode(args: ClaimVotingRewardsArgs): Hex {
+  return encodeAbiParameters(claimVotingRewardsEncoding, [
+    args.allTokensToTransfer,
+    args.claimLockerRewards,
+    args.extraRewardTokens,
+    args.votiumClaims,
+    args.unstakeCvxCrv,
+  ]);
+}
+
+export function claimVotingRewardsDecode(encoded: Hex): ClaimVotingRewardsArgs {
+  const [allTokensToTransfer, claimLockerRewards, extraRewardTokens, votiumClaims, unstakeCvxCrv] = decodeAbiParameters(
+    claimVotingRewardsEncoding,
+    encoded,
+  );
+
+  return {
+    allTokensToTransfer,
+    claimLockerRewards,
+    extraRewardTokens,
+    votiumClaims,
+    unstakeCvxCrv,
+  };
+}
+
+//--------------------------------------------------------------------------------------------
+// DELEGATE
+//--------------------------------------------------------------------------------------------
+
+export const delegate = ExternalPositionManager.makeUse(Action.Delegate, delegateEncode);
+export const createAndDelegate = ExternalPositionManager.makeCreateAndUse(Action.Delegate, delegateEncode);
+
+const delegateEncoding = [
+  {
+    name: "delegate",
+    type: "address",
+  },
+] as const;
+
+export type DelegateArgs = {
+  delegate: Address;
+};
+
+export function delegateEncode(args: DelegateArgs): Hex {
+  return encodeAbiParameters(delegateEncoding, [args.delegate]);
+}
+
+export function delegateDecode(encoded: Hex): DelegateArgs {
+  const [delegate] = decodeAbiParameters(delegateEncoding, encoded);
+
+  return {
+    delegate,
+  };
+}
+
+//--------------------------------------------------------------------------------------------
+// EXTERNAL READ FUNCTIONS - CVX MINING
 //--------------------------------------------------------------------------------------------
 
 const cvxMiningAbi = {
@@ -30,7 +448,7 @@ export async function convertCrvToCvx(
 }
 
 //--------------------------------------------------------------------------------------------
-// VOTE LOCKED CONVEX TOKEN
+// EXTERNAL READ FUNCTIONS - VOTE LOCKED CONVEX TOKEN
 //--------------------------------------------------------------------------------------------
 
 // ABI from https://etherscan.io/address/0x72a19342e8f1838460ebfccef09f6585e32db86e#code
@@ -151,7 +569,7 @@ export async function getUserLocks(
 }
 
 //--------------------------------------------------------------------------------------------
-// CVX BOOSTER
+// EXTERNAL READ FUNCTIONS - CVX BOOSTER
 //--------------------------------------------------------------------------------------------
 
 // ABI from https://etherscan.io/address/0xf403c135812408bfbe8713b5a23a04b3d48aae31#code
@@ -279,7 +697,7 @@ export async function getPlatformFee(
 }
 
 //--------------------------------------------------------------------------------------------
-// CVX CRV REWARDS
+// EXTERNAL READ FUNCTIONS - CVX CRV REWARDS
 //--------------------------------------------------------------------------------------------
 
 // ABI from https://etherscan.io/address/0x3fe65692bfcd0e6cf84cb1e7d24108e434a7587e#code
@@ -352,7 +770,7 @@ export async function getExtraRewards(
 }
 
 //--------------------------------------------------------------------------------------------
-// CVX CRV EXTRA REWARDS
+// EXTERNAL READ FUNCTIONS - CVX CRV EXTRA REWARDS
 //--------------------------------------------------------------------------------------------
 
 export async function getExtraRewardsRewards(
@@ -384,7 +802,7 @@ export async function getExtraRewardsRewardToken(
 }
 
 //--------------------------------------------------------------------------------------------
-// STAKING WRAPPER
+// EXTERNAL READ FUNCTIONS - STAKING WRAPPER
 //--------------------------------------------------------------------------------------------
 
 export async function getEstimateRewards(
