+++ packages/sdk/src/Portfolio/Integrations/Kiln.ts
@@ -1,5 +1,5 @@
-import { type Address, type Hex, decodeAbiParameters, encodeAbiParameters } from "viem";
-import { Assertion } from "../../Utils.js";
+import { type Address, type Hex, PublicClient, decodeAbiParameters, encodeAbiParameters, parseAbi } from "viem";
+import { Assertion, Viem } from "../../Utils.js";
 import * as ExternalPositionManager from "../../_internal/ExternalPositionManager.js";
 
 export type Action = typeof Action[keyof typeof Action];
@@ -155,4 +155,51 @@ export function unstakeDecode(encoded: Hex): UnstakeArgs {
     stakingContract,
     publicKeys,
   };
+}
+
+//--------------------------------------------------------------------------------------------
+// EXTERNAL READ FUNCTIONS
+//--------------------------------------------------------------------------------------------
+
+export async function getCLFeeRecipient(
+  client: PublicClient,
+  args: Viem.ContractCallParameters<{
+    kilnStaking: Address;
+    validatorPublicKey: Hex;
+  }>,
+) {
+  return Viem.readContract(client, args, {
+    abi: parseAbi(["function getCLFeeRecipient(bytes calldata _publicKey) external view returns (address)"]),
+    functionName: "getCLFeeRecipient",
+    address: args.kilnStaking,
+    args: [args.validatorPublicKey],
+  });
+}
+
+export async function getELFeeRecipient(
+  client: PublicClient,
+  args: Viem.ContractCallParameters<{
+    kilnStaking: Address;
+    validatorPublicKey: Hex;
+  }>,
+) {
+  return Viem.readContract(client, args, {
+    abi: parseAbi(["function getELFeeRecipient(bytes calldata _publicKey) external view returns (address)"]),
+    functionName: "getELFeeRecipient",
+    address: args.kilnStaking,
+    args: [args.validatorPublicKey],
+  });
+}
+
+export async function getGlobalFee(
+  client: PublicClient,
+  args: Viem.ContractCallParameters<{
+    kilnStaking: Address;
+  }>,
+) {
+  return Viem.readContract(client, args, {
+    abi: parseAbi(["function getGlobalFee() external view returns (uint256 globalFee_)"]),
+    functionName: "getGlobalFee",
+    address: args.kilnStaking,
+  });
 }
