+++ packages/sdk/src/Portfolio/Integrations/TheGraph.ts
@@ -1,4 +1,5 @@
-import { type Address, type Hex, decodeAbiParameters, encodeAbiParameters } from "viem";
+import { type Address, type Hex, PublicClient, decodeAbiParameters, encodeAbiParameters, parseAbi } from "viem";
+import { Viem } from "../../Utils.js";
 import * as ExternalPositionManager from "../../_internal/ExternalPositionManager.js";
 
 export type Action = typeof Action[keyof typeof Action];
@@ -114,4 +115,57 @@ export function withdrawDecode(encoded: Hex): WithdrawArgs {
     indexer,
     nextIndexer,
   };
+}
+
+//--------------------------------------------------------------------------------------------
+// EXTERNAL READ FUNCTIONS
+//--------------------------------------------------------------------------------------------
+
+export async function getDelegationPool(
+  client: PublicClient,
+  args: Viem.ContractCallParameters<{
+    stakingContract: Address;
+    indexer: Address;
+  }>,
+) {
+  const [cooldownBlocks, indexingRewardCut, queryFeeCut, updatedAtBlock, tokens, shares] = await Viem.readContract(
+    client,
+    args,
+    {
+      abi: parseAbi([
+        "function delegationPools(address indexer) view returns (uint32 cooldownBlocks, uint32 indexingRewardCut, uint32 queryFeeCut, uint256 updatedAtBlock, uint256 tokens, uint256 shares)",
+      ]),
+      functionName: "delegationPools",
+      address: args.stakingContract,
+      args: [args.indexer],
+    },
+  );
+
+  return { cooldownBlocks, indexingRewardCut, queryFeeCut, updatedAtBlock, tokens, shares };
+}
+
+export function getDelegationTaxPercentage(
+  client: PublicClient,
+  args: Viem.ContractCallParameters<{
+    stakingContract: Address;
+  }>,
+) {
+  return Viem.readContract(client, args, {
+    abi: parseAbi(["function delegationTaxPercentage() view returns (uint32)"]),
+    functionName: "delegationTaxPercentage",
+    address: args.stakingContract,
+  });
+}
+
+export function getCurrentEpoch(
+  client: PublicClient,
+  args: Viem.ContractCallParameters<{
+    epochManager: Address;
+  }>,
+) {
+  return Viem.readContract(client, args, {
+    abi: parseAbi(["function currentEpoch() view returns (uint256)"]),
+    functionName: "currentEpoch",
+    address: args.epochManager,
+  });
 }
